Convert markdown or rmarkdown files to mindmap files.

Convert markdown or rmarkdown files to mindmap files.

md2mm(pattern = "*.[R]*md$", title = NA, path = ".", remove_curly_bracket = FALSE,
    savefile = TRUE, savefilename = NA, backup = TRUE, bookdown_style = TRUE,
    keep_eq = FALSE, method = c("regexpr", "pandoc"))

Arguments

pattern
an optional regular expression for filtering the input files. See help(dir).

title
character. The title of the output file.

path
character. The path of the folder which contains the input file(s).

remove_curly_bracket
logical. Whether to remove #ID in the headers of the markdown file (usually in a 'bookdown' https://github.com/rstudio/bookdown project).

savefile
logical. Whether to save the output as a file.

savefilename
character. Valid when savefile == TRUE.

backup
logical. Whether the existing target file, if any, should be saved as backup.

bookdown_style
logical. whether the markdown files are in bookdown style, i.e. index.Rmd at the beginning, # (PART), # (APPENDIX) and # References as an upper level of normal # title

keep_eq
logical. whether to keep LaTeX equations.

method
"regexpr" uses regular expressions, 'pandoc' uses pandoc to find the headings.

Value

a mindmap file, which can be viewed by common mindmap software, such as 'FreeMind' (http://freemind.sourceforge.net/wiki/index.php/Main_Page) and 'XMind' (http://www.xmind.net).

Examples

path <- system.file("examples/md", package = "mindr")
md2mm(path = path)

#> md.mm was generated!

#>  [1] "<map version=\"1.0.1\">"                                                                                                  
#>  [2] "<node TEXT=\"md\">"                                                                                                       
#>  [3] "<node TEXT=\"Introduction {#introduction}\">"                                                                             
#>  [4] "<node TEXT=\"What is 'bookdown' {#what_is__bookdown_}\"></node>"                                                          
#>  [5] "<node TEXT=\"What is 'bookdownplus'\" LINK=\"https://github.com/pzhaonet/bookdownplus {#what_is__bookdownplus_}\"></node>"
#>  [6] "<node TEXT=\"Why 'bookdownplus' {#why__bookdownplus_}\"></node>"                                                          
#>  [7] "<node TEXT=\"Giants' Shoulders {#giants__shoulders}\"></node></node>"                                                     
#>  [8] "<node TEXT=\"Quick Start {#quick_start}\">"                                                                               
#>  [9] "<node TEXT=\"Preparation {#preparation}\"></node>"                                                                        
#> [10] "<node TEXT=\"Installation of 'bookdownplus' {#installation_of__bookdownplus_}\"></node>"                                  
#> [11] "<node TEXT=\"How to use {#how_to_use}\"></node>"                                                                          
#> [12] "<node TEXT=\"More output formats\"></node>"                                                                               
#> [13] "<node TEXT=\"More templates\"></node>"                                                                                    
#> [14] "<node TEXT=\"A magic trick\"></node>"                                                                                     
#> [15] "<node TEXT=\"Recommendations\"></node></node>"                                                                            
#> [16] "<node TEXT=\"Basic {#basic}\">"                                                                                           
#> [17] "<node TEXT=\"Markdown Syntax {#markdown_syntax}\">"                                                                       
#> [18] "<node TEXT=\"What is Markdown {#what_is_markdown}\"></node>"                                                              
#> [19] "<node TEXT=\"Basic syntax {#basic_syntax}\"></node>"                                                                      
#> [20] "<node TEXT=\"Chapters {#chapters}\"></node>"                                                                              
#> [21] "<node TEXT=\"Figures and tables {#figures_and_tables}\"></node>"                                                          
#> [22] "<node TEXT=\"References {#references}\"></node>"                                                                          
#> [23] "<node TEXT=\"Theorems, lemma, definitions, etc. {#theorems__lemma__definitions__etc_}\"></node>"                          
#> [24] "<node TEXT=\"Export Word document {#export_word_document}\"></node>"                                                      
#> [25] "<node TEXT=\"Equations numbering {#equations_numbering}\"></node></node>"                                                 
#> [26] "<node TEXT=\"R, RStudio and bookdown {#r__rstudio_and_bookdown}\"></node>"                                                
#> [27] "<node TEXT=\"LaTeX and Pandoc {#latex_and_pandoc}\"></node>"                                                              
#> [28] "<node TEXT=\"Workflow {#workflow}\"></node></node>"                                                                       
#> [29] "<node TEXT=\"Simple {#simple}\"></node>"                                                                                  
#> [30] "<node TEXT=\"Lifestyle {#lifestyle}\">"                                                                                   
#> [31] "<node TEXT=\"Journal {#journal}\"></node>"                                                                                
#> [32] "<node TEXT=\"Poem book {#poem_book}\"></node>"                                                                            
#> [33] "<node TEXT=\"Music {#music}\"></node></node>"                                                                             
#> [34] "<node TEXT=\"Office {#office}\">"                                                                                         
#> [35] "<node TEXT=\"Mail {#mail}\">"                                                                                             
#> [36] "<node TEXT=\"Arguments for mail content\"></node>"                                                                        
#> [37] "<node TEXT=\"Mail themes\"></node></node>"                                                                                
#> [38] "<node TEXT=\"Calendar {#calendar}\"></node></node>"                                                                       
#> [39] "<node TEXT=\"Academic {#academic}\">"                                                                                     
#> [40] "<node TEXT=\"Articles {#articles}\"></node>"                                                                              
#> [41] "<node TEXT=\"Thesis {#thesis}\"></node>"                                                                                  
#> [42] "<node TEXT=\"Poster {#poster}\"></node>"                                                                                  
#> [43] "<node TEXT=\"Chemistry {#chemistry}\"></node></node>"                                                                     
#> [44] "<node TEXT=\"Advanced {#advanced}\">"                                                                                     
#> [45] "<node TEXT=\"Chinese {#chinese}\"></node>"                                                                                
#> [46] "<node TEXT=\"Mind Map {#mind_map}\"></node>"                                                                              
#> [47] "<node TEXT=\"Create Your Own Templates {#customize}\"></node></node>"                                                     
#> [48] "<node TEXT=\"FAQ {#faq}\"></node>"                                                                                        
#> [49] "<node TEXT=\"Bibliography {-}\"></node>"                                                                                  
#> [50] "</node></map>"                                                                                                            

md2mm(path = path, remove_curly_bracket = TRUE)

#> md-2019-01-16-11-58-30.mm was generated!

#>  [1] "<map version=\"1.0.1\">"                                                                         
#>  [2] "<node TEXT=\"md\">"                                                                              
#>  [3] "<node TEXT=\"Introduction \">"                                                                   
#>  [4] "<node TEXT=\"What is 'bookdown' \"></node>"                                                      
#>  [5] "<node TEXT=\"What is 'bookdownplus'\" LINK=\"https://github.com/pzhaonet/bookdownplus \"></node>"
#>  [6] "<node TEXT=\"Why 'bookdownplus' \"></node>"                                                      
#>  [7] "<node TEXT=\"Giants' Shoulders \"></node></node>"                                                
#>  [8] "<node TEXT=\"Quick Start \">"                                                                    
#>  [9] "<node TEXT=\"Preparation \"></node>"                                                             
#> [10] "<node TEXT=\"Installation of 'bookdownplus' \"></node>"                                          
#> [11] "<node TEXT=\"How to use \"></node>"                                                              
#> [12] "<node TEXT=\"More output formats\"></node>"                                                      
#> [13] "<node TEXT=\"More templates\"></node>"                                                           
#> [14] "<node TEXT=\"A magic trick\"></node>"                                                            
#> [15] "<node TEXT=\"Recommendations\"></node></node>"                                                   
#> [16] "<node TEXT=\"Basic \">"                                                                          
#> [17] "<node TEXT=\"Markdown Syntax \">"                                                                
#> [18] "<node TEXT=\"What is Markdown \"></node>"                                                        
#> [19] "<node TEXT=\"Basic syntax \"></node>"                                                            
#> [20] "<node TEXT=\"Chapters \"></node>"                                                                
#> [21] "<node TEXT=\"Figures and tables \"></node>"                                                      
#> [22] "<node TEXT=\"References \"></node>"                                                              
#> [23] "<node TEXT=\"Theorems, lemma, definitions, etc. \"></node>"                                      
#> [24] "<node TEXT=\"Export Word document \"></node>"                                                    
#> [25] "<node TEXT=\"Equations numbering \"></node></node>"                                              
#> [26] "<node TEXT=\"R, RStudio and bookdown \"></node>"                                                 
#> [27] "<node TEXT=\"LaTeX and Pandoc \"></node>"                                                        
#> [28] "<node TEXT=\"Workflow \"></node></node>"                                                         
#> [29] "<node TEXT=\"Simple \"></node>"                                                                  
#> [30] "<node TEXT=\"Lifestyle \">"                                                                      
#> [31] "<node TEXT=\"Journal \"></node>"                                                                 
#> [32] "<node TEXT=\"Poem book \"></node>"                                                               
#> [33] "<node TEXT=\"Music \"></node></node>"                                                            
#> [34] "<node TEXT=\"Office \">"                                                                         
#> [35] "<node TEXT=\"Mail \">"                                                                           
#> [36] "<node TEXT=\"Arguments for mail content\"></node>"                                               
#> [37] "<node TEXT=\"Mail themes\"></node></node>"                                                       
#> [38] "<node TEXT=\"Calendar \"></node></node>"                                                         
#> [39] "<node TEXT=\"Academic \">"                                                                       
#> [40] "<node TEXT=\"Articles \"></node>"                                                                
#> [41] "<node TEXT=\"Thesis \"></node>"                                                                  
#> [42] "<node TEXT=\"Poster \"></node>"                                                                  
#> [43] "<node TEXT=\"Chemistry \"></node></node>"                                                        
#> [44] "<node TEXT=\"Advanced \">"                                                                       
#> [45] "<node TEXT=\"Chinese \"></node>"                                                                 
#> [46] "<node TEXT=\"Mind Map \"></node>"                                                                
#> [47] "<node TEXT=\"Create Your Own Templates \"></node></node>"                                        
#> [48] "<node TEXT=\"FAQ \"></node>"                                                                     
#> [49] "<node TEXT=\"Bibliography \"></node>"                                                            
#> [50] "</node></map>"